functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.4.0
google-cloud-firestore>=2.14.0
